requirements.txt
aiogram==3.4.1
alembic==1.13.1
APScheduler==3.10.4
asyncpg==0.29.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28